requirements.txt
django
pathlib
channels
asgi_redis
RPi.GPIO